Removed duplication between BeakerQueryCell and ChatQueryCell.
